+++ common/src/Model/BrushFace.cpp
@@ -25,11 +25,11 @@
 #include "Exceptions.h"
 #include "FloatType.h"
 #include "Model/MapFormat.h"
-#include "Model/ParallelTexCoordSystem.h"
-#include "Model/ParaxialTexCoordSystem.h"
+#include "Model/ParallelUVCoordSystem.h"
+#include "Model/ParaxialUVCoordSystem.h"
 #include "Model/TagMatcher.h"
 #include "Model/TagVisitor.h"
-#include "Model/TexCoordSystem.h"
+#include "Model/UVCoordSystem.h"
 #include "Polyhedron.h"
 
 #include "kdl/overload.h"
