+++ common/src/Model/UVCoordSystem.cpp
@@ -17,11 +17,11 @@
  along with TrenchBroom. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "TexCoordSystem.h"
+#include "UVCoordSystem.h"
 
 #include "Model/BrushFace.h"
-#include "Model/ParallelTexCoordSystem.h"
-#include "Model/ParaxialTexCoordSystem.h"
+#include "Model/ParallelUVCoordSystem.h"
+#include "Model/ParaxialUVCoordSystem.h"
 
 #include "vm/mat.h"
 #include "vm/mat_ext.h"
