+++ common/test/src/Model/tst_UVCoordSystem.cpp
@@ -20,8 +20,8 @@
 #include "Assets/Material.h"
 #include "FloatType.h"
 #include "Model/BrushFaceAttributes.h"
-#include "Model/ParallelTexCoordSystem.h"
-#include "Model/ParaxialTexCoordSystem.h"
+#include "Model/ParallelUVCoordSystem.h"
+#include "Model/ParaxialUVCoordSystem.h"
 
 #include "vm/vec.h"
