+++ common/test/src/TestUtils.cpp
@@ -30,8 +30,8 @@
 #include "Model/EntityNode.h"
 #include "Model/GameImpl.h"
 #include "Model/GroupNode.h"
-#include "Model/ParallelTexCoordSystem.h"
-#include "Model/ParaxialTexCoordSystem.h"
+#include "Model/ParallelUVCoordSystem.h"
+#include "Model/ParaxialUVCoordSystem.h"
 #include "Model/PatchNode.h"
 #include "Model/WorldNode.h"
 #include "TestLogger.h"
